+++ frontend/app/course/[courseId]/page.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { getServerSession } from 'next-auth/next'
-import { authOptions } from '../../api/auth/[...nextauth]/route'
 
-import { getCourseData } from '../../api/course/courseData'
-import { CourseDetail } from '../../../components/pages/CourseDetail'
+import { authOptions } from '../../api/auth/[...nextauth]/route'
+import { CourseWithSectionsType } from '../../../types/CourseType'
+import { CourseDetailWrapper } from '../../../components/wrapper/pages/CourseDetail'
 import { QuestionType } from '../../../types/QuestionType'
 import { AnswerType } from '../../../types/AnswerType'
 import Error from '../../error'
@@ -16,9 +16,19 @@ export default async function CourseDetailPage({
   searchParams: { videoId: string; questionId: string }
 }) {
   const session = await getServerSession(authOptions)
+  const userId = session?.user?.id
 
   try {
-    const initialCourseData = await getCourseData(params.courseId)
+    const getCourseData = await fetch(
+      `${process.env.NEXT_PUBLIC_FRONT_API_URL}/course/${params.courseId}`,
+      {
+        cache: 'no-cache',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      },
+    )
+    const initialCourseData: CourseWithSectionsType = await getCourseData.json()
 
     const getQuestionsData = await fetch(
       `${process.env.NEXT_PUBLIC_BACKEND_URL}/question/${searchParams.videoId}`,
@@ -43,10 +53,12 @@ export default async function CourseDetailPage({
     const answers: AnswerType[] = await getAnswersData.json()
 
     return (
-      <CourseDetail
+      <CourseDetailWrapper
         courseId={params.courseId}
         initialCourseData={initialCourseData}
         session={session}
+        userId={userId}
+        videoId={searchParams.videoId}
         questions={questions}
         answers={answers}
         questionId={searchParams.questionId}
